+++ src/game_data/other_constants.rs
@@ -1,6 +1,6 @@
-pub const LEVEL_WIDTH: usize = 1600;
-pub const LEVEL_HEIGHT: usize = 160;
-pub const VGASPEC_POSITION: usize = 304;
+pub const LEVEL_WIDTH: u32 = 1600;
+pub const LEVEL_HEIGHT: u32 = 160;
+pub const VGASPEC_POSITION: u32 = 304;
 
-pub const SCREEN_WIDTH: usize = 320;
-pub const SCREEN_HEIGHT: usize = 200;
+pub const SCREEN_WIDTH: u32 = 320;
+pub const SCREEN_HEIGHT: u32 = 200;
