+++ src/game_data/skill_panel.rs
@@ -1,32 +1,31 @@
-use crate::game_data::{Bitmap, SCREEN_WIDTH, Sprite};
-
 pub use crate::game_data::file::main::{
     resolve_skill_panel_font_index, resolve_skill_panel_skill_font_index,
 };
+use crate::game_data::{Bitmap, SCREEN_WIDTH, Sprite};
 
-pub const SKILL_PANEL_HEIGHT: usize = 40;
+pub const SKILL_PANEL_HEIGHT: u32 = 40;
 
-pub const MINIMAP_FRAME_HEIGHT: usize = 20;
-pub const MINIMAP_FRAME_WIDTH: usize = 22;
+pub const MINIMAP_FRAME_HEIGHT: u32 = 20;
+pub const MINIMAP_FRAME_WIDTH: u32 = 22;
 const FRAME_COLOR: u8 = 3;
 
-pub const MINIMAP_AREA_WIDTH: usize = 104;
-pub const MINIMAP_AREA_HEIGHT: usize = 20;
-pub const MINIMAP_AREA_X: usize = 208;
-pub const MINIMAP_AREA_Y: usize = 18;
+pub const MINIMAP_AREA_WIDTH: u32 = 104;
+pub const MINIMAP_AREA_HEIGHT: u32 = 20;
+pub const MINIMAP_AREA_X: u32 = 208;
+pub const MINIMAP_AREA_Y: u32 = 18;
 
-pub const MINIMAP_VIEW_WIDTH: usize = 100;
-pub const MINIMAP_VIEW_HEIGHT: usize = 18;
-pub const MINIMAP_VIEW_X: usize = MINIMAP_AREA_X + 2;
-pub const MINIMAP_VIEW_Y: usize = MINIMAP_AREA_Y + 1;
+pub const MINIMAP_VIEW_WIDTH: u32 = 100;
+pub const MINIMAP_VIEW_HEIGHT: u32 = 18;
+pub const MINIMAP_VIEW_X: u32 = MINIMAP_AREA_X + 2;
+pub const MINIMAP_VIEW_Y: u32 = MINIMAP_AREA_Y + 1;
 
-pub const SKILL_TILE_WIDTH: usize = 16;
-pub const SKILL_TILE_HEIGHT: usize = 24;
-pub const SKILL_TILE_Y: usize = 16;
-pub const SKILL_TILE_LABEL_X: usize = 4;
-pub const SKILL_TILE_LABEL_Y: usize = 1;
-pub const SKILL_TILE_LABEL_WIDTH: usize = 8;
-pub const SKILL_TILE_LABEL_HEIGHT: usize = 8;
+pub const SKILL_TILE_WIDTH: u32 = 16;
+pub const SKILL_TILE_HEIGHT: u32 = 24;
+pub const SKILL_TILE_Y: u32 = 16;
+pub const SKILL_TILE_LABEL_X: u32 = 4;
+pub const SKILL_TILE_LABEL_Y: u32 = 1;
+pub const SKILL_TILE_LABEL_WIDTH: u32 = 8;
+pub const SKILL_TILE_LABEL_HEIGHT: u32 = 8;
 
 #[derive(Clone)]
 pub struct SkillPanel {
@@ -41,14 +40,17 @@ impl SkillPanel {
     pub fn new(mut panel: Bitmap, font: Sprite, font_skills: Sprite) -> SkillPanel {
         for y in MINIMAP_AREA_Y..MINIMAP_AREA_Y + MINIMAP_AREA_HEIGHT {
             for x in MINIMAP_AREA_X..MINIMAP_AREA_X + MINIMAP_AREA_WIDTH {
-                panel.transparency[x + y * SCREEN_WIDTH] = true;
+                panel.transparency[(x + y * SCREEN_WIDTH) as usize] = true;
             }
         }
 
         SkillPanel {
             panel,
-            minimap_frame: create_frame(MINIMAP_FRAME_WIDTH, MINIMAP_FRAME_HEIGHT),
-            skill_tile_frame: create_frame(SKILL_TILE_WIDTH, SKILL_PANEL_HEIGHT),
+            minimap_frame: create_frame(
+                MINIMAP_FRAME_WIDTH as usize,
+                MINIMAP_FRAME_HEIGHT as usize,
+            ),
+            skill_tile_frame: create_frame(SKILL_TILE_WIDTH as usize, SKILL_PANEL_HEIGHT as usize),
             font,
             font_skills,
         }
